refactor(sidebar): extract default-branch cache, move test helper out of prod bundle
Split `useDefaultBranch` + its module-level cache into
`default-branch-cache.ts`, and move `__resetDefaultBranchCacheForTests`
into a colocated `sidebar-workspace-row.test-utils.ts` that is only
imported from test files. Tree-shaking now drops the test helper from
production bundles entirely, closing the leak flagged in the Phase 3
audit.
